data_pipeline/requirements.txt
# Core dependencies
google-cloud-storage>=2.10.0
google-cloud-pubsub
numpy>=1.24.0
pandas>=2.0.0
supabase>=1.0.3
python-dotenv>=1.0.0
filelock>=3.12.2
spacy==3.5.3
huggingface-hub>=0.19.0
psycopg2-binary>=2.9.0
nltk>=3.8.1
scikit-learn>=1.0.0
apache-airflow>=2.10.0
apache-airflow-providers-google>=10.0.0
torch
pymupdf
mammoth
xgboost
imblearn